Pull tremor docs as part of build process
Stores tremor docs repo as a git submodule. The repo is then used as part
of the language server build process, for generating the function documentation
as well as module completion items. Once we store those items in a structured
way as part of the tremor-script codebase, this won't be needed.

This also enables us to publish the language server to crates.io, since the
build does not depend on manual clone of the docs repo now.