Fix cargo publish and install for v0.8
For v0.8, we introduced tremor-runtime as a git submodule dependency but
during the crate publish process, it was found that tremor-runtime
files were not actually included in the crate archive (in contrast to
tremor-www-docs, which has been a successful dependency since v0.7).
This inclusion failure seems to be because of the size of tremor-runtime
repo, resulting in build failures during `cargo publish` process.

With `cargo publish --no-verify`, we can publish the crate. However,
installation from this published crate fails since our current build
script relies on being run from a git repo (for submodule dependency)
and the pulled crate is not a git repo.

To work around this, we now try to read the tremor-runtime files (mainly
tremor-script stdlib) from the local cargo registry where tremor-script
is pulled as part of the trill build. If that fails for some reason, we
sparately download the tremor-script crate file (corresponding to the
version in Cargo.lock dependency) and work from there.

This now allows us to eliminate the (heavy) tremor-runtime git submodule
dependency completely. The version of tremor-script stdlib now read for
trill's doc features (on hover/completion) are also now automatically
in sync with the tremor-script version specified as trill's dependency.